Use abstract base class for UBXNode and UORBNode to reduce amount of code to maintain
